Naming updates

+++ McEliece.Cryptosystem.Security/McEliece.Cryptosystem.Security/Models/AverageEntry.cs
@@ -5,7 +5,7 @@
         public int L0SetCount { get; set; }
         public int Count { get; set; }
         public int Sum { get; set; }
-        public int Average { get; set; }
+        public double Average { get; set; }
         public int L0SetCountOccurrences { get; set; }
     }
 }
